WIP: credentials-binding implementation for Docker client certificates
This is an implementation of Binding (the Extension for credentials-binding plugin) for DockerServerCredentials (client certificates). It gives a simple way to make these credentials available in any kind of build step or pipeline jobs, using the usual credentials-binding mecanism. The {ca,cert,key}.pem files will be written in a temporary directory, whose path can be bound to (by default) $DOCKER_CERT_PATH.

This PR depends on a little refactoring in credentials-binding-plugin, to reuse code from their FileBinding implementation - see jenkinsci/credentials-binding-plugin#25. Until it is accepted or rejected, this PR can't be merged, so I'm only submitting it for comments/reference.