Adapt DockerServerCredentialsTest for 2.413 UI improvement
Jenkins 2.412 and earlier use a text area for the client key.

Jenkins 2.413 and newer use an input field for the client key.

Jenkins 2.413 looks much nicer with a button that is consistent with
the other buttons used in the Jenkins pages.

https://github.com/jenkinsci/jenkins/pull/8179 is the Jenkins core pull
request that implemented the change in Jenkins 2.413.  Other plugins
that use HTMLUnit to enter a value for the secret text may need a
similar change.